### 2025-06-03 #8214 @Morgane-des-Ligneris
Hi @arfon,
Thanks for the update. Based on the similar JOSS papers highlighted, here are a few potential reviewers who I believe could be a good fit :
- ollyk, author of Volume Segmentics, a tool for volumetric segmentation using deep learning, with relevant experience in medical image segmentation.
- hkjeldsberg, author of VaMPy, a pipeline for modeling vascular geometries, with relevant expertise in vessel segmentation and analysis.
- rmdocherty, author of SAMBA, which involves segmentation and annotation tools, conceptually aligned with my submission.
- pchlap, author of PlatiPy, a medical imaging toolkit in Python, with relevant experience in processing workflows and analysis tools.

Please let me know if any further details are needed.
Thanks again.
