### 2025-06-13 c3e18e3
Fix contour_points retrieval to handle missing data
