### 2025-06-14 1233825
Update graph_branches.py
FIX: removed useless emptiness double check
