### 2025-06-16 f01e790
Revert "Fix contour_points retrieval to handle missing data"
This reverts commit c3e18e3af712eb0d49765c01e0fc3d9a4113ddb9.
