### 2025-07-06 #8214 @dstansby
Hi @Morgane-des-Ligneris, I'll be the editor for your submission - sorry for the wait. It's my first time as editor, so you might have to bear with me a little bit, but I'll try my best to make the process as smooth as possible!

The sumbmission looks in scope and good to send to review to me, so I'll look for reviewers now - thanks for the reviewer suggestions above.
