### 2025-07-09 #8563 @editorialbot
[REVIEW]: VSMOD: A Vessel Segmentation and MODelization plugin for 3D Slicer
**Submitting author:** @Morgane-des-Ligneris (<a href="http://orcid.org/0000-0003-1008-3009">Morgane des Ligneris</a>)
**Repository:** https://github.com/omerveille/VSMOD
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
