### 2025-07-11 omerveille/PulmonaryArteriesSegmentor#12 @dstansby
Can't install plugin from source
Despite the clear instructions, I can't seem to install the plugin successfully. Steps I've taken:

1. Clone this repository
2. Install 3D Slicer (5.8.1 r33241 / 11eaf62, on macOS)
3. Go to Developer Tools > Extension Wizard
4. Click "Select Extension"
5. Select the folder for the locally cloned repository

At this point I get a view on the left like this:

but I did not get a popup as is described in this step:

> A popup will show, select Yes.⚠️ Make sure you are connected to the internet. When loading the module, it will check if you have all the dependencies and attempt to download the missing ones. ⚠️

After following the steps above, VSMOD doesn't appear in the Segmentation menu.
