### 2025-07-14 omerveille/VSMOD#12 @dstansby
Thanks for the clear instructions! I followed them, and Slicer re-opened fine. There is indeed an errror in the Python console though (see below). It looks like `intel-openmp` is only available for x86_64 Linux and Windows, and not macOS. So I guess this means that VSMOD is not compatible with macOS? For my own testing I have access to a Windows computer, so I can try on that (using the branch you provided).

```python
Installing missing dependencies...
Installing numpy...
Collecting numpy>=2.0.2
  Downloading numpy-2.0.2-cp39-cp39-macosx_14_0_x86_64.whl.metadata (60 kB)

Downloading numpy-2.0.2-cp39-cp39-macosx_14_0_x86_64.whl (6.9 MB)

Installing collected packages: numpy
  Attempting uninstall: numpy
    Found existing installation: numpy 1.26.4
    Uninstalling numpy-1.26.4:
      Successfully uninstalled numpy-1.26.4
Successfully installed numpy-2.0.2
numpy installed !
Installing trimesh...
Collecting trimesh>=4.6.5
  Downloading trimesh-4.7.0-py3-none-any.whl.metadata (18 kB)
Requirement already satisfied: numpy>=1.20 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from trimesh>=4.6.5) (2.0.2)
Downloading trimesh-4.7.0-py3-none-any.whl (708 kB)

Installing collected packages: trimesh
Successfully installed trimesh-4.7.0
trimesh installed !
Installing scikit-image...
Collecting scikit-image>=0.24.0
  Downloading scikit_image-0.24.0-cp39-cp39-macosx_10_9_x86_64.whl.metadata (14 kB)
Requirement already satisfied: numpy>=1.23 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from scikit-image>=0.24.0) (2.0.2)
Requirement already satisfied: scipy>=1.9 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from scikit-image>=0.24.0) (1.13.1)
Collecting networkx>=2.8 (from scikit-image>=0.24.0)
  Downloading networkx-3.2.1-py3-none-any.whl.metadata (5.2 kB)
Requirement already satisfied: pillow>=9.1 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from scikit-image>=0.24.0) (10.3.0)
Collecting imageio>=2.33 (from scikit-image>=0.24.0)
  Using cached imageio-2.37.0-py3-none-any.whl.metadata (5.2 kB)
Collecting tifffile>=2022.8.12 (from scikit-image>=0.24.0)
  Downloading tifffile-2024.8.30-py3-none-any.whl.metadata (31 kB)
Requirement already satisfied: packaging>=21 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from scikit-image>=0.24.0) (24.0)
Collecting lazy-loader>=0.4 (from scikit-image>=0.24.0)
  Using cached lazy_loader-0.4-py3-none-any.whl.metadata (7.6 kB)
Downloading scikit_image-0.24.0-cp39-cp39-macosx_10_9_x86_64.whl (14.1 MB)

Using cached imageio-2.37.0-py3-none-any.whl (315 kB)
Using cached lazy_loader-0.4-py3-none-any.whl (12 kB)
Downloading networkx-3.2.1-py3-none-any.whl (1.6 MB)

Downloading tifffile-2024.8.30-py3-none-any.whl (227 kB)

Installing collected packages: tifffile, networkx, lazy-loader, imageio, scikit-image
Successfully installed imageio-2.37.0 lazy-loader-0.4 networkx-3.2.1 scikit-image-0.24.0 tifffile-2024.8.30
scikit-image installed !
Installing networkx...
Requirement already satisfied: networkx>=3.2.1 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (3.2.1)
networkx installed !
Installing numba...
Collecting numba>=0.60.0
  Downloading numba-0.60.0-cp39-cp39-macosx_10_9_x86_64.whl.metadata (2.7 kB)
Collecting llvmlite<0.44,>=0.43.0dev0 (from numba>=0.60.0)
  Downloading llvmlite-0.43.0-cp39-cp39-macosx_10_9_x86_64.whl.metadata (4.8 kB)
Requirement already satisfied: numpy<2.1,>=1.22 in /Applications/Slicer.app/Contents/lib/Python/lib/python3.9/site-packages (from numba>=0.60.0) (2.0.2)
Downloading numba-0.60.0-cp39-cp39-macosx_10_9_x86_64.whl (2.6 MB)

Downloading llvmlite-0.43.0-cp39-cp39-macosx_10_9_x86_64.whl (31.1 MB)

Installing collected packages: llvmlite, numba
Successfully installed llvmlite-0.43.0 numba-0.60.0
numba installed !
Installing intel-openmp...
ERROR: Could not find a version that satisfies the requirement intel-openmp>=2024.2.1 (from versions: 2018.0.0, 2018.0.3, 2019.0, 2021.1.1, 2021.1.2, 2021.2.0, 2021.3.0, 2021.4.0, 2022.0.0, 2022.1.0, 2022.2.0, 2022.2.1, 2023.0.0, 2023.1.0, 2023.2.0)
ERROR: No matching distribution found for intel-openmp>=2024.2.1
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/Applications/Slicer.app/Contents/lib/Python/lib/python3.9/imp.py", line 169, in load_source
    module = _exec(spec, sys.modules[name])
  File "<frozen importlib._bootstrap>", line 613, in _exec
  File "<frozen importlib._bootstrap_external>", line 850, in exec_module
  File "<frozen importlib._bootstrap>", line 228, in _call_with_frames_removed
  File "/Users/dstansby/software/joss/VSMOD/VSMOD/VSMOD.py", line 47, in <module>
    from ransac_slicer.dependencies_checker import check_and_install_missing_dependencies
  File "/Users/dstansby/software/joss/VSMOD/VSMOD/ransac_slicer/__init__.py", line 4, in <module>
    restart_needed = check_and_install_missing_dependencies()
  File "/Users/dstansby/software/joss/VSMOD/VSMOD/ransac_slicer/dependencies_checker.py", line 118, in check_and_install_missing_dependencies
    install_missing_module(missing_modules)
  File "/Users/dstansby/software/joss/VSMOD/VSMOD/ransac_slicer/dependencies_checker.py", line 91, in install_missing_module
    pip_install(command)
  File "/Applications/Slicer.app/Contents/bin/Python/slicer/util.py", line 3942, in pip_install
    _executePythonModule("pip", args)
  File "/Applications/Slicer.app/Contents/bin/Python/slicer/util.py", line 3896, in _executePythonModule
    logProcessOutput(proc)
  File "/Applications/Slicer.app/Contents/bin/Python/slicer/util.py", line 3862, in logProcessOutput
    raise CalledProcessError(retcode, proc.args, output=proc.stdout, stderr=proc.stderr)
subprocess.CalledProcessError: Command '['/Applications/Slicer.app/Contents/bin/../bin/PythonSlicer', '-m', 'pip', 'install', 'intel-openmp>=2024.2.1']' returned non-zero exit status 1.
[Python] Failed to install dependencies.
[Python] Command '['/Applications/Slicer.app/Contents/bin/../bin/PythonSlicer', '-m', 'pip', 'install', 'intel-openmp>=2024.2.1']' returned non-zero exit status 1.
[Qt] loadSourceAsModule - Failed to load file "/Users/dstansby/software/joss/VSMOD/VSMOD/VSMOD.py"  as module "VSMOD" !
[Qt] Fail to instantiate module  "VSMOD"
[Qt] The following modules failed to be instantiated:
[Qt]    VSMOD
```
