### 2025-07-16 fa3d219
fix: changes to allow plugin compatibility on  macOS, encoding and threading layers issues made it incompatible
