### 2025-07-16 omerveille/VSMOD#12 @Leirbag-gabrieL
Hello, I am back with some updates !

Indeed, It seems like `intel-openmp` is not available for macOS via pip.
`intel-openmp` is used by the threading layer of  `numba` in order to parallelize some crucial portions of the codes.

From the testing we have done with @Azeline, it seems like `tbb`  can't serve as an alternative threading layer for `numba` because Slicer's python version is outdated and `numba` requires a newer version of `tbb`.

So only for macOS, we will use the built-in `numba` threading layer, which may be slightly slower but work reliably.
We did not find yet why 3D slicer do not add our module to the modules paths, but it seems to be a known issue on macOS linked to the naming of the files.

I have pushed a fix on this branch: https://github.com/omerveille/VSMOD/tree/os_specific_changes, it should work on macOS but you will still have to do the “manual” install I explained in my previous message.

Thank you for your detailed feedback, it’s been incredibly helpful for debugging! Let me know if you encounter further issues.
