### 2025-07-18 9a56b40
fix: renamed 'VSMOD' to 'vsmod' to make 3D Slicer add the module to module paths
