+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
   rev: v0.1.3
   hooks:
     - id: ruff
-      args: ["--fix", "--show-fixes",  "--line-length=120", "--extend-exclude", "VSMOD/module_tests/__init__.py"]
+      args: ["--fix", "--show-fixes",  "--line-length=120", "--extend-exclude", "vsmod/module_tests/__init__.py"]
     - id: ruff-format
 
 - repo: https://github.com/asottile/pyupgrade
