+++ vsmod/CMakeLists.txt
@@ -1,5 +1,5 @@
 #-----------------------------------------------------------------------------
-set(MODULE_NAME VSMOD)
+set(MODULE_NAME vsmod)
 
 #-----------------------------------------------------------------------------
 set(MODULE_PYTHON_SCRIPTS
