### 2025-07-29 #8563 @OllyK
> 👋 [@Morgane-des-Ligneris](https://github.com/Morgane-des-Ligneris); I've started my review, and have one question and one problem:
>
> 1. Looking at the commit history of the plugin, it looks like you have not significantly contributed to the code. Could you please explain your relation to the code? We don't have hard and fast rules on authorship, but just want to make sure it's appropriate.
> 2. I'm having issues installing the plugin to continue my review; I opened an issue at [Can't install plugin from source omerveille/VSMOD#12](https://github.com/omerveille/VSMOD/issues/12), hopefully we can resolve my issues there so I can get started trying out the plugin.

Hi @Morgane-des-Ligneris 👋 I'm also starting my review now. Thanks for your description of your role in authoring this software - Along the same topics as dstansby's comments I would like to ask:

1.  How did the three last authors contribute to this project? As stated in the JOSS guidelines, authorship is a complex topic and there are many different conventions.

2. Thanks to Gabriel Husak for the fix for installation of the plugin on the Mac. Now that the installation procedure for Macs differs from the standard install, would it be possible to update the installation instructions accordingly? In addition, since 3D Slicer has an extension manager that allows for easy searching and installation of plugins I would like to ask if you had considered adding VSMOD as an official extension? It could simplify installation for non-technical users. To clarify I am not requesting this to be done as part of this review, it is a suggested improvement to make things easier in the future.
