### 2025-08-18 403b998
Clarify description of RVXLiverSegmentation plugin's applicability to liver vasculature in paper.md
