+++ Paper/paper.md
@@ -52,7 +52,7 @@ Accurate annotation of vascular networks in medical imaging remains a critical c
 
 Existing vessel annotation tools have notable limitations. Traditional paint and brush tools in image analysis software such as *3D Slicer*[^1] [@pinter2019] or *ImageJ*[^2] required slice-by-slice pixel-level annotation, making the process extremely laborious and often resulting in disconnected segmentations. Tubular shape prior tools have been proposed, such as the *Draw Tube* tool from the [SlicerSegmentEditorExtraEffects extension](https://github.com/lassoan/SlicerSegmentEditorExtraEffects). However, this method requires manually drawing each vessel separately, and its fixed tubular shape is poorly suited for real vessels, which exhibit tortuosity and variable radii. Additionally, constructing a fully connected vascular tree remains particularly challenging.
 
-Recently, Lamy *et al.* proposed the *RVXLiverSegmentation* plugin [@lamy2022] for 3D Slicer to segment hepatic vascular networks. This tool enables fast and accurate vascular segmentation by requiring users to place predetermined points at branching nodes. However, it is designed specifically for the hepatic vasculature and is not applicable to other organs.
+Recently, Lamy *et al.* proposed the *RVXLiverSegmentation* plugin [@lamy2022] for 3D Slicer to segment vascular networks of the liver. This tool enables fast and accurate vascular segmentation by requiring users to place predetermined points at branching nodes. However, it is designed specifically for the vasculature of the liver and is not applicable to other organs.
 *VSMOD* is designed to overcome these limitations by providing a user-friendly plugin for efficiently generating accurate and topologically correct vascular segmentations.
 
 [^1]: [https://www.slicer.org/](https://www.slicer.org/)
