### 2025-08-18 3e9cd22
Add macOS installation instructions and notes to README.md
