### 2025-08-18 0e35dfb
Add testing section to README.md with instructions for running automated tests
