+++ Paper/paper.bib
@@ -141,3 +141,56 @@
   keywords = {annotation quality,benchmark,cerebrovascular segmentation,semi-supervised learning},
   file = {/home/desligneris/snap/zotero-snap/common/Zotero/storage/LJFRLVV5/Rougé et al. - 2025 - Guidelines for cerebrovascular segmentation Managing imperfect annotations in the context of semi-s.pdf}
 }
+
+@InProceedings{esteves2025,
+  author="Carneiro-Esteves, Sophie and Vacavant, Antoine and Merveille, Odyss{\'e}e",
+  editor="Chen, Chao and Singh, Yash and Hu, Xiaoling",
+  title="Restoring Connectivity in Vascular Segmentations Using a Learned Post-processing Model",
+  booktitle="Topology- and Graph-Informed Imaging Informatics",
+  year="2025",
+  publisher="Springer Nature Switzerland",
+  address="Cham",
+  pages="55--65",
+  abstract="Accurate segmentation of vascular networks is essential for computer-aided tools designed to address cardiovascular diseases. Despite more than thirty years of research, it remains a challenge to obtain vascular segmentation results that preserve the connectivity of the underlying vascular network. Yet connectivity is one of the key features of these tools. In this work, we propose a post-processing algorithm aiming to reconnect vascular structures that have been disconnected by a segmentation algorithm. Connectivity being a complex property to model explicitly, we propose to learn this geometric feature either through synthetic data or annotations of the application of interest. The resulting post-processing model can be used on the output of any supervised or unsupervised vascular segmentation algorithm. We show that this post-processing effectively restores the connectivity of vascular networks both in 2D and 3D images, leading to improved overall segmentation results.",
+  isbn="978-3-031-73967-5"
+}
+
+@article{rouge2024,
+    doi = {10.1371/journal.pone.0311439},
+    author = {Rougé, Pierre AND Passat, Nicolas AND Merveille, Odyssée},
+    journal = {PLOS ONE},
+    publisher = {Public Library of Science},
+    title = {Topology aware multitask cascaded U-Net for cerebrovascular segmentation},
+    year = {2024},
+    month = {12},
+    volume = {19},
+    url = {https://doi.org/10.1371/journal.pone.0311439},
+    pages = {1-20},
[…]
