+++ Paper/paper.md
@@ -48,7 +48,7 @@ This plugin significantly reduces the time required to obtain complete volumetri
 
 
 # Statement of need
-Accurate annotation of vascular networks in medical imaging remains a critical challenge and is essential to develop accurate deep learning-based segmentation models. Current segmentation models still struggle with preserving the connectivity of complex vascular networks which is crucial for downstream tasks such as flow dynamic simulations. Improving the quality and precision of annotated datasets is crucial for training more effective models. However, traditional manual annotation methods are extremely time-consuming and subject to inter-observer variability creating important concept shifts in the annotation, ultimately leading to lower model performance [@rouge2025].
+Accurate annotation of vascular networks in medical imaging remains a critical challenge and is essential to develop accurate deep learning-based segmentation models. Current segmentation models still struggle with preserving the connectivity of complex vascular networks [@esteves2025; @rouge2024; @Keshwani2020] which is crucial for downstream tasks such as flow dynamic simulations. Improving the quality and precision of annotated datasets is crucial for training more effective models. However, traditional manual annotation methods are extremely time-consuming and subject to inter-observer variability creating important concept shifts in the annotation, ultimately leading to lower model performance [@rouge2025].
 
 Existing vessel annotation tools have notable limitations. Traditional paint and brush tools in image analysis software such as *3D Slicer*[^1] [@pinter2019] or *ImageJ*[^2] required slice-by-slice pixel-level annotation, making the process extremely laborious and often resulting in disconnected segmentations. Tubular shape prior tools have been proposed, such as the *Draw Tube* tool from the [SlicerSegmentEditorExtraEffects extension](https://github.com/lassoan/SlicerSegmentEditorExtraEffects). However, this method requires manually drawing each vessel separately, and its fixed tubular shape is poorly suited for real vessels, which exhibit tortuosity and variable radii. Additionally, constructing a fully connected vascular tree remains particularly challenging.
 
