### 2025-08-25 #8563 @Morgane-des-Ligneris
Dear [OllyK](https://github.com/OllyK),

Thank you for your review and constructive feedback. I apologize for the delay in my response, I was on holiday for the past few weeks.

1. Regarding the contributions of the last three authors:
	- Guillaume Dollé is the original author of the RANSAC algorithm code that we adapted for this project. He provided valuable technical guidance on implementing the algorithm for our specific use case, and also contributed feedback and revisions to improve the paper.
	- Carole Frindel and Odyssée Merveille are my thesis supervisors who were actively involved throughout the project. They were involved during the entire development process, provided scientific guidance and feedback, and contributed significantly to reviewing and improving the paper.
2. Regarding installation instructions and extension manager: Thank you for this suggestion. I updated the installation instructions to reflect the Mac-specific installation procedure that Gabriel provided. You are absolutely right about the 3D Slicer extension manager - our ultimate goal is indeed to make VSMOD an official extension available directly through 3D Slicer's extension manager, which would greatly simplify installation for non-technical users.

Please let me know if you need any additional information.
