### 2025-08-25 #8563 @Morgane-des-Ligneris
Dear [dstansby](https://github.com/dstansby),

Thank you for your positive feedback about the software! I addressed your three points:

1. Automated testing instructions: I added comprehensive instructions on running automated tests to the README. I also included the important detail that 3D Slicer needs to be set to developer mode to access the "Reload and Test" button.
2. Installation instructions: As mentioned to [OllyK](https://github.com/OllyK), I updated these to address the installation issues.
3. Citation for connectivity preservation challenges: I included literature references to support this statement about current segmentation models struggling with vascular connectivity preservation.

Thank you for suggesting replacing "hepatic" with "of the liver" - this is a good point for improving accessibility, and I have made this change. Please let me know if you need any additional information.
