### 2025-09-01 omerveille/PulmonaryArteriesSegmentor#17 @OllyK
Suggested edits for JOSS paper submission
Hi @Morgane-des-Ligneris
With reference to your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/8563), I am creating this issue with some minor suggested edits to the paper manuscript just some typos and things which may make it clearer to read.

*line 12* - to simplify and lighten -> to simplify and **streamline**
*line 18* - tracking new branch -> tracking new branch**es**/ **a** new branch?
*line 71* - the algorithm the first time -> the algorithm **for** the first time
line 93 - Ten computed tomography pulmonary angiography (CTPA) were extensively -> Ten computed tomography pulmonary angiography (CTPA) **images/datasets** were extensively
*line 94* - the proposed plugin by an expert which reported -> the **VSMOD** plugin by an expert **who** reported
*line 96* - the centerline module, where users manually placed -> the centerline module, where **the user** manually placed/the centerline module, where users manually **place** ?
Legend for Table 1: Comparison of average segmentation time between manual annotation and VSMOD for a patient with an average of 269 vessels -> Comparison of average segmentation times **per** patient **calculated on data with an** average of 269 vessels
*Line 114* - and cerebral vessels, VSMOD performed well, as illustrated in -> and cerebral vessels, VSMOD **was also able to segment these vessels**, as illustrated in
*Line 115* - VSMOD provides a user-friendly framework -> **In conclusion**, VSMOD provides a user-friendly framework

Send me a comment if you need some clarification on any of these,

Olly
