### 2025-09-01 #8563 @OllyK
Hi @Morgane-des-Ligneris
Thanks for getting back to me with clarification and I hope you had a good holiday. I've read through the manuscript now, overall it reads well and is a great peice of work. As you'll see I've created [an issue](https://github.com/omerveille/VSMOD/issues/17) with suggestions for minor alterations - just some things which may make it clearer to read.

Best Wishes,
Olly
