### 2025-09-01 omerveille/PulmonaryArteriesSegmentor#18 @OllyK
Error when running automated tests on macOS
This issue is created as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8563)
When running the automated suite of module tests on macOS, the following error is output:

```python
ERROR: test_01_time_function_and_flush_timers (module_tests.helper_test.HelperTest)
Traceback (most recent call last):
  File "/Users/olly/Code/repos/VSMOD/vsmod/module_tests/helper_test.py", line 41, in test_01_time_function_and_flush_timers
    flushed_file_path = flush_timers()
  File "/Users/olly/Code/repos/VSMOD/vsmod/ransac_slicer/helper.py", line 35, in flush_timers
    timer_file_path.touch(exist_ok=True)
  File "/Applications/Slicer.app/Contents/lib/Python/lib/python3.9/pathlib.py", line 1315, in touch
    fd = self._raw_open(flags, mode)
  File "/Applications/Slicer.app/Contents/lib/Python/lib/python3.9/pathlib.py", line 1127, in _raw_open
    return self._accessor.open(self, flags, mode)
OSError: [Errno 30] Read-only file system: "timers_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.json"
```
It looks like an attempt is made to write a file to the current working directory of the Slicer.app, which is a read-only on macOS.

There are two minor fixes that could be made to the [`flush_timers()` function](https://github.com/omerveille/VSMOD/blob/f1b28983cec83367d5570f06bd3cdf4346bf3e4e/vsmod/ransac_slicer/helper.py#L24) to remedy this:

1. Use the `tempfile` library to use the system temp directory instead of the working directory and prepend this to the path for the JSON file.
2. The [f-string](https://github.com/omerveille/VSMOD/blob/f1b28983cec83367d5570f06bd3cdf4346bf3e4e/vsmod/ransac_slicer/helper.py#L33) for the output JSON file seems to be incorrectly formatted and is missing the "f", so the output filename will be incorrect.

e.g.
```python
import tempfile
from pathlib import Path
from datetime import datetime

def flush_timers() -> Path:
    """
    Write timers to disk near slicer executable.

    Returns:
        The path to the created file.
    """

    timer_file_path = Path(tempfile.gettempdir()) / f"timers_{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}.json"
    timer_file_path.touch(exist_ok=True)
    ...
```
Thanks,

Olly
