### 2025-09-01 omerveille/VSMOD#19 @OllyK
Add link to sample data file to use during tutorial
As part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8563) it would be useful to have a direct link to download some (small) data that could be used when working through the documentation on how to use your plugin. At the moment, a person using the plugin and getting started, already has to have data of their own, else has to use one of the example datasets from the 3DSlicer app. It would be nice, if possible, to have access to the same data that is used in your example videos.
