### 2025-09-01 #8563 @OllyK
Hi again @Morgane-des-Ligneris
I have now finished my review and would like to congratulate you and your collegues on creating a nice piece of software with associated documentation. There are three things like I would like to see fixed:
1.  Some minor edits to the manuscript - [Issue here](https://github.com/omerveille/VSMOD/issues/17)
2. A fix to the automated tests on macOS - [Issue here](https://github.com/omerveille/VSMOD/issues/18)
3. A link (or some other description) allowing access to a suitable data file to load when working through the tutorial documentation - [Issue here](https://github.com/omerveille/VSMOD/issues/19)
