### 2025-09-02 omerveille/VSMOD#19 @Leirbag-gabrieL
I have added a “sample data” section in the README with a link to an open-source dataset (IXI in this case). I believe the image used in the video is not publicly available.

The full IXI MRA dataset is quite large (11.5 GB), so I recommend testing with this [image](https://github.com/omerveille/VSMOD/blob/main/vsmod/Resources/Test/IXI002-Guys-0828-MRA.nii.gz)
, which is also used in the automated tests. If you specifically want to use the same image from the video, you could ask @Morgane-des-Ligneris or @omerveille  for permission.
