### 2025-09-03 omerveille/VSMOD#19 @OllyK
Hi @Leirbag-gabrieL ,

Thanks for updating the README, would it be possible to put the link in to the testing image that you posted above? It seems to be a good one to use rather than dealing with the full IXI MRA dataset,

Thanks,

Olly
