### 2025-09-08 ee9b04b
Re-use acronym
