+++ Paper/paper.bib
@@ -152,12 +152,13 @@
   address="Cham",
   pages="55--65",
   abstract="Accurate segmentation of vascular networks is essential for computer-aided tools designed to address cardiovascular diseases. Despite more than thirty years of research, it remains a challenge to obtain vascular segmentation results that preserve the connectivity of the underlying vascular network. Yet connectivity is one of the key features of these tools. In this work, we propose a post-processing algorithm aiming to reconnect vascular structures that have been disconnected by a segmentation algorithm. Connectivity being a complex property to model explicitly, we propose to learn this geometric feature either through synthetic data or annotations of the application of interest. The resulting post-processing model can be used on the output of any supervised or unsupervised vascular segmentation algorithm. We show that this post-processing effectively restores the connectivity of vascular networks both in 2D and 3D images, leading to improved overall segmentation results.",
-  isbn="978-3-031-73967-5"
+  isbn="978-3-031-73967-5",
+  doi={10.1007/978-3-031-73967-5_6}
 }
 
 @article{rouge2024,
     doi = {10.1371/journal.pone.0311439},
-    author = {Rougé, Pierre AND Passat, Nicolas AND Merveille, Odyssée},
+    author = {Rougé, Pierre and Passat, Nicolas and Merveille, Odyssée},
     journal = {PLOS ONE},
     publisher = {Public Library of Science},
     title = {Topology aware multitask cascaded U-Net for cerebrovascular segmentation},
@@ -192,5 +193,6 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="14--23",
 abstract="Reconstructing Portal Vein and Hepatic Vein trees from contrast enhanced abdominal CT scans is a prerequisite for preoperative liver surgery simulation. Existing deep learning based methods treat vascular tree reconstruction as a semantic segmentation problem. However, vessels such as hepatic and portal vein look very similar locally and need to be traced to their source for robust label assignment. Therefore, semantic segmentation by looking at local 3D patch results in noisy misclassifications. To tackle this, we propose a novel multi-task deep learning architecture for vessel tree reconstruction. The network architecture simultaneously solves the task of detecting voxels on vascular centerlines (i.e. nodes) and estimates connectivity between center-voxels (edges) in the tree structure to be reconstructed. Further, we propose a novel connectivity metric which considers both inter-class distance and intra-class topological distance between center-voxel pairs. Vascular trees are reconstructed starting from the vessel source using the learned connectivity metric using the shortest path tree algorithm. A thorough evaluation on public IRCAD dataset shows that the proposed method considerably outperforms existing semantic segmentation based methods. To the best of our knowledge, this is the first deep learning based approach which learns multi-label tree structure connectivity from images.",
-isbn="978-3-030-59725-2"
+isbn="978-3-030-59725-2",
+doi={10.1007/978-3-030-59725-2_2}
 }
