### 2025-09-29 #8563 @Morgane-des-Ligneris
Hi [dstansby](https://github.com/dstansby),
Everything looks good for the paper. Is there one last step required that I am missing before the accept command with the editorial bot is used?
Thank you again !
