### 2025-10-08 #8563 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] author list of archive should match author list of paper
- [ ] make sure affiliations include countries (some have only cities at the moment)
- [ ] The paper is on the long side for JOSS; we generally target 1000 words or less. Much of the detail regarding algorithm operation is likely more appropriate for documentation – I'll request you to take a shot at compressing things down there, and after that I can do a copyediting pass to finalize things.
