### 2025-10-08 8df593a
Added missing countries in affiliations and simplified detail regarding algorithm operation in module descriptions in paper.md
