### 2025-10-08 #8563 @Morgane-des-Ligneris
Hello [rkurchin](https://github.com/rkurchin),
Sorry for the confusion in the author list on the zenodo archive, it is now fixed, as well as the country in each affiliations.
For the length of the paper, we updated the content in the module description, and it is now around 1200 words total. If it is still too much, I think there is a way to simplify the result part. Let me know what you think, and I will update the zenodo archive afterwards.
