### 2025-10-09 1affc26
Fix punctuation in segmentation module description
